Metadata Fallback + Rewriting Blog Section (See Desc)
The twitter logo metadata field now has a fallback to the duck image.

Also, rewriting the blog section:
Implemented an owl carousel selector for category and painstakingly did
all the javascript and css for it

Implementing (in liquid) loops which output html layouts of posts in the
blog section. Currently has sections of 3,2 and 1. (The css for 3 is
only fully complete)

In the future:
- Make multiple layout formats for different number of posts
- Choose layouts randomly for variety
- Do css for the post layouts
- More layouts like for 6 posts